requirements.txt
fastapi
uvicorn
python-hcl2
chromadb
google-generativeai
python-dotenv
PyGithub
tiktoken
